// app/models/User.tsx
export interface User {
    id: number | undefined;
    firstName: string | undefined;
    name: string | undefined;
    email: string | undefined;
    password: string | undefined;
    role: string | undefined;
    createdAt: string | undefined;
    updatedAt: string | undefined;
    avatar: string | undefined;
}